src/main/python/airflow/extract/gcp_fetch/requirements.txt
google-cloud-storage>=1.42.0
requests>=2.26.0
bs4==0.0.2
beautifulsoup4==4.12.3
google-cloud-secret-manager>=2.7.0
python-dotenv==1.0.0
